+++ include/parse_frame.h
@@ -15,8 +15,8 @@
  * Logs one parse frame
  */
 void pf_log(
-   log_sev_t     logsev, /**< [in]  */
-   parse_frame_t *pf     /**< [in]  */
+   log_sev_t      logsev, /**< [in]  */
+   parse_frame_t* pf      /**< [in]  */
 );
 
 
@@ -24,8 +24,8 @@ void pf_log(
  * Copies src to dst.
  */
 void pf_copy(
-   parse_frame_t       *dst, /**< [in]  */
-   const parse_frame_t *src  /**< [in]  */
+   parse_frame_t*       dst, /**< [in]  */
+   const parse_frame_t* src  /**< [in]  */
 );
 
 
@@ -34,7 +34,7 @@ void pf_copy(
  * This is called on #if and #ifdef.
  */
 void pf_push(
-   parse_frame_t *pf /**< [in]  */
+   parse_frame_t* pf /**< [in]  */
 );
 
 
@@ -44,7 +44,7 @@ void pf_push(
  * This is called on the first #else and #elif.
  */
 void pf_push_under(
-   parse_frame_t *pf /**< [in]  */
+   parse_frame_t* pf /**< [in]  */
 );
 
 
@@ -53,7 +53,7 @@ void pf_push_under(
  * This is called on #else and #elif.
  */
 void pf_copy_tos(
-   parse_frame_t *pf /**< [in]  */
+   parse_frame_t* pf /**< [in]  */
 );
 
 
@@ -68,16 +68,16 @@ void pf_trash_tos(void);
  * This is called on #endif
  */
 void pf_pop(
-   parse_frame_t *pf /**< [in]  */
+   parse_frame_t* pf /**< [in]  */
 );
 
 
 /**
  * Returns the pp_indent to use for this line
  */
-size_t pf_check(
-   parse_frame_t *frm,/**< [in]  */
-   chunk_t       *pc  /**< [in]  */
+uint32_t pf_check(
+   parse_frame_t* frm,/**< [in]  */
+   chunk_t*       pc  /**< [in]  */
 );
